Se cambio la matriz de distancias de tipo int a double

+++ include/MST.h
@@ -8,12 +8,12 @@
 
 using namespace std;
  
-int *primMST(int n,int p /* # de ojbetivos */,int ***graph);
-void printMST(int n,int *parent, int **graph);
+int *primMST(int n,int p /* # de ojbetivos */,double ***graph);
+void printMST(int n,int *parent, double **graph);
 int testMST();
-int ***read_instance (int *n,int *p);
-void getMST(int *n,int *p,int ***graph);
-void delete_graph(int n,int p,int ***graph);
+double ***read_instance (int *n,int *p);
+void getMST(int *n,int *p,double ***graph);
+void delete_graph(int n,int p,double ***graph);
 
 #endif
